Restructure GitHub provider files and update imports
