+++ internal-packages/workflow-designer-ui/src/editor/data-source/provider/github/install-application.tsx
@@ -1,12 +1,14 @@
 import { useIntegration } from "@giselle-sdk/integration/react";
 import { useCallback, useEffect, useTransition } from "react";
-import { GitHubIcon, SpinnerIcon } from "../../../icons";
-import { usePopupWindow } from "../../lib/use-popup-window";
+import { GitHubIcon, SpinnerIcon } from "../../../../icons";
+import { usePopupWindow } from "../../../lib/use-popup-window";
 
-export function Unauthorized({ authUrl }: { authUrl: string }) {
-	const { refresh } = useIntegration();
+export function InstallGitHubApplication({
+	installationUrl,
+}: { installationUrl: string }) {
 	const [isPending, startTransition] = useTransition();
-	const { open } = usePopupWindow(authUrl);
+	const { refresh } = useIntegration();
+	const { open } = usePopupWindow(installationUrl);
 
 	const handleInstallationMessage = useCallback(
 		(event: MessageEvent) => {
@@ -29,7 +31,10 @@ export function Unauthorized({ authUrl }: { authUrl: string }) {
 	return (
 		<div className="bg-white-900/10 h-[300px] rounded-[8px] flex items-center justify-center">
 			<div className="flex flex-col gap-[8px]">
-				<p>To get started you have to sign into your GitHub account</p>
+				<p>
+					Install the GitHub application for the accounts you with to Import
+					from to continue
+				</p>
 				<button
 					type="button"
 					className="group cursor-pointer bg-black-900 rounded-[4px] py-[4px] flex items-center justify-center gap-[8px] disabled:opacity-50 disabled:cursor-wait"
@@ -37,7 +42,7 @@ export function Unauthorized({ authUrl }: { authUrl: string }) {
 					disabled={isPending}
 				>
 					<GitHubIcon className="size-[18px]" />
-					Continue with GitHub
+					Install
 					<SpinnerIcon className="hidden group-disabled:block animate-follow-through-overlap-spin" />
 				</button>
 			</div>
